+++ src/kryptos_hash_common.c
@@ -33,7 +33,7 @@ void kryptos_hash_apply_pad_on_u32_block(kryptos_u32_t *input, size_t const inpu
     }
 }
 
-void kryptos_hash_ld_u8buf_as_u32_blocks(kryptos_u8_t *buffer, const int buffer_size,
+void kryptos_hash_ld_u8buf_as_u32_blocks(kryptos_u8_t *buffer, const size_t buffer_size,
                                          kryptos_u32_t *input, const size_t input_nr,
                                          const size_t *block_index_decision_table) {
     size_t b, i;
@@ -70,7 +70,7 @@ void kryptos_hash_apply_pad_on_u64_block(kryptos_u64_t *input, size_t const inpu
     }
 }
 
-void kryptos_hash_ld_u8buf_as_u64_blocks(kryptos_u8_t *buffer, const int buffer_size,
+void kryptos_hash_ld_u8buf_as_u64_blocks(kryptos_u8_t *buffer, const size_t buffer_size,
                                          kryptos_u64_t *input, const size_t input_nr,
                                          const size_t *block_index_decision_table) {
     size_t b, i;
